backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
google-generativeai==0.5.4
openai==1.30.1
tiktoken==0.7.0
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1